02-tool-use/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
anthropic==0.52.0
pydantic==2.11.1
python-dotenv==1.1.0
httpx==0.28.1
